% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvap.R
\name{fvap_context}
\alias{fvap_context}
\title{Return context: full source of this client}
\usage{
fvap_context()
}
\value{
character: source code of this file
}
\description{
Return context: full source of this client
}
