% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvap.R
\name{fvap_summary}
\alias{fvap_summary}
\title{Summarize an FVAP survey dataset (column names and types)}
\usage{
fvap_summary(dataset)
}
\arguments{
\item{dataset}{Dataset id from fvap_list()}
}
\value{
tibble: column_name, column_type, n_missing, n_unique
}
\description{
Summarize an FVAP survey dataset (column names and types)
}
